+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 1.2.0
+
+- Added badges:
+	- participation;
+	- subject.
+- Added participation in `student-index`.
+- Changed honorable color from gray to blue.
+- Removed medal count in `student-index`.
+
 # 1.1.0
 
 - Sort by generation on `student-index`.